refactor(scripts): extract generate-comprehensive-uat-tests.js into domain modules
SD-LEO-REFAC-UAT-GEN-003: Domain extraction refactoring

Original file: 896 LOC
After refactoring:
- generate-comprehensive-uat-tests.js: 34 LOC (thin wrapper)
- generate-uat/index.js: 180 LOC (main orchestrator)
- generate-uat/test-suite-config.js: 268 LOC
- generate-uat/test-generators.js: 386 LOC
- generate-uat/suite-builders.js: 166 LOC

All modules under 400 LOC (target: 500)
Backward compatibility maintained via re-exports
