+++ scripts/generate-comprehensive-uat-tests.js
@@ -3,894 +3,32 @@
 /**
  * Comprehensive UAT Test Suite Generator for EHG Application
  * Generates 28 test files covering all 43+ pages with ~330 test cases
+ *
+ * REFACTORED: This file is now a thin wrapper around the domain modules.
+ * See scripts/generate-uat/ for the extracted domain architecture.
+ *
+ * Domains:
+ * - test-suite-config.js: Test suite definitions and configurations
+ * - test-generators.js: Test file and implementation generators
+ * - suite-builders.js: Admin, cross-functional, E2E suite builders
+ * - index.js: Main orchestrator with re-exports
  */
 
-import fs from 'fs/promises';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import { dirname } from 'path';
-import chalk from 'chalk';
+// Re-export everything from the domain modules for backward compatibility
+export {
+  generateAllTests,
+  TEST_SUITES,
+  generateTestFile,
+  generateTestImplementation,
+  generateAdminTests,
+  generateCrossFunctionalTests,
+  generateE2ETests
+} from './generate-uat/index.js';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
+export { default } from './generate-uat/index.js';
 
-const TEST_DIR = path.join(__dirname, '..', 'tests', 'uat');
-const BASE_URL = 'http://localhost:8080';
-
-/**
- * Test Suite Configuration
- * Each suite includes multiple test scenarios for broad and deep coverage
- */
-const TEST_SUITES = {
-  // Phase 1: Core User Journeys
-  ventures: {
-    name: 'Venture Management',
-    route: '/ventures',
-    tests: [
-      'List all ventures with pagination',
-      'Create new venture - basic flow',
-      'Create new venture - with all optional fields',
-      'Edit venture details - name and description',
-      'Edit venture details - financial data',
-      'Edit venture details - team members',
-      'Delete venture with confirmation',
-      'Cancel delete operation',
-      'Search ventures by name',
-      'Search ventures by status',
-      'Filter ventures by category',
-      'Filter ventures by date range',
-      'Sort ventures by name',
-      'Sort ventures by value',
-      'Sort ventures by created date',
-      'Bulk select ventures',
-      'Bulk delete ventures',
-      'Bulk export ventures',
-      'View venture details',
-      'Navigate between ventures',
-      'Add venture to portfolio',
-      'Remove venture from portfolio',
-      'Share venture report',
-      'Print venture summary',
-      'Venture quick actions menu',
-      'Venture status transitions',
-      'Venture permission checks',
-      'Venture data validation',
-      'Venture audit trail',
-      'Venture performance metrics'
-    ]
-  },
-
-  dashboard: {
-    name: 'Executive Dashboard',
-    route: '/chairman',
-    tests: [
-      'Dashboard initial load performance',
-      'All widgets render correctly',
-      'Real-time metrics update',
-      'Interactive chart hover states',
-      'Chart drill-down functionality',
-      'Widget refresh buttons',
-      'Dashboard layout customization',
-      'Save dashboard preferences',
-      'Export dashboard as PDF',
-      'Export dashboard data as CSV',
-      'Quick action buttons work',
-      'Notification panel displays',
-      'Notification mark as read',
-      'Clear all notifications',
-      'KPI cards show correct data',
-      'Performance indicators update',
-      'Date range selector works',
-      'Dashboard search functionality',
-      'Recent activity feed',
-      'Upcoming events calendar',
-      'Task list management',
-      'Dashboard responsive layout',
-      'Dark mode toggle',
-      'Accessibility keyboard navigation',
-      'Dashboard help tooltips'
-    ]
-  },
-
-  analytics: {
-    name: 'Analytics & Reporting',
-    route: '/analytics',
-    tests: [
-      'Analytics dashboard loads',
-      'Report templates available',
-      'Generate standard report',
-      'Generate custom report',
-      'Report preview functionality',
-      'Export report as PDF',
-      'Export report as Excel',
-      'Email report to recipients',
-      'Schedule recurring reports',
-      'Data visualization options',
-      'Chart type switching',
-      'Apply data filters',
-      'Date range selection',
-      'Comparison periods',
-      'Trend analysis tools',
-      'Predictive analytics',
-      'Custom metrics builder',
-      'Save report template',
-      'Report sharing permissions',
-      'Analytics performance benchmarks'
-    ]
-  },
-
-  aiAgents: {
-    name: 'AI Agent Management',
-    route: '/ai-agents',
-    tests: [
-      'AI agents list page loads',
-      'CEO Agent activation',
-      'CEO Agent task delegation',
-      'CEO Agent response handling',
-      'GTM Strategist activation',
-      'GTM strategy generation',
-      'GTM plan approval flow',
-      'Competitive Intelligence agent',
-      'Competitor analysis report',
-      'Market insights generation',
-      'Creative Media agent activation',
-      'Content generation workflow',
-      'Content approval process',
-      'Agent coordination dashboard',
-      'Multi-agent collaboration',
-      'Agent task history',
-      'Agent performance metrics',
-      'Agent configuration settings',
-      'Agent permission management',
-      'Agent error handling',
-      'Agent fallback scenarios',
-      'Agent learning feedback',
-      'Agent API integration',
-      'Agent webhook handling',
-      'Agent notification system'
-    ]
-  },
-
-  eva: {
-    name: 'EVA Assistant',
-    route: '/eva-orchestration',
-    tests: [
-      'EVA chat interface loads',
-      'Send text message to EVA',
-      'Receive EVA response',
-      'EVA command execution',
-      'EVA context awareness',
-      'EVA multi-turn conversation',
-      'EVA suggestion chips',
-      'EVA quick actions',
-      'EVA history retrieval',
-      'Clear conversation history',
-      'EVA file upload handling',
-      'EVA data analysis',
-      'EVA report generation',
-      'EVA task automation',
-      'EVA integration commands',
-      'EVA help system',
-      'EVA error recovery',
-      'EVA session persistence',
-      'EVA voice input',
-      'EVA export conversation'
-    ]
-  },
-
-  workflows: {
-    name: 'Workflow Management',
-    route: '/workflows',
-    tests: [
-      'Workflows list page loads',
-      'Create new workflow',
-      'Edit workflow steps',
-      'Delete workflow',
-      'Workflow templates',
-      'Workflow execution start',
-      'Workflow progress tracking',
-      'Workflow pause/resume',
-      'Workflow cancellation',
-      'Workflow completion',
-      'Workflow error handling',
-      'Workflow notifications',
-      'Workflow approval chains',
-      'Workflow automation rules',
-      'Workflow performance metrics'
-    ]
-  },
-
-  portfolios: {
-    name: 'Portfolio Management',
-    route: '/portfolios',
-    tests: [
-      'Portfolio overview loads',
-      'Create new portfolio',
-      'Add assets to portfolio',
-      'Remove assets from portfolio',
-      'Portfolio performance charts',
-      'Portfolio allocation view',
-      'Portfolio risk assessment',
-      'Portfolio rebalancing tool',
-      'Portfolio comparison',
-      'Portfolio reports',
-      'Portfolio sharing',
-      'Portfolio notifications',
-      'Portfolio settings',
-      'Portfolio archival',
-      'Portfolio restoration'
-    ]
-  },
-
-  governance: {
-    name: 'Governance & Compliance',
-    route: '/governance',
-    tests: [
-      'Governance dashboard loads',
-      'Policy list displays',
-      'Create new policy',
-      'Edit existing policy',
-      'Policy approval workflow',
-      'Compliance check execution',
-      'Audit trail viewing',
-      'Generate compliance report',
-      'Risk assessment tools',
-      'Governance notifications'
-    ]
-  },
-
-  integrations: {
-    name: 'External Integrations',
-    route: '/integrations',
-    tests: [
-      'Integration hub loads',
-      'View available integrations',
-      'Configure new integration',
-      'Test integration connection',
-      'Integration data sync',
-      'Webhook configuration',
-      'API key management',
-      'Integration logs viewing',
-      'Integration error handling',
-      'Disable integration'
-    ]
-  },
-
-  team: {
-    name: 'Team & Collaboration',
-    route: '/team',
-    tests: [
-      'Team page loads',
-      'View team members',
-      'Invite new member',
-      'Edit member permissions',
-      'Remove team member',
-      'Team activity feed',
-      'Team performance metrics',
-      'Team communication tools',
-      'Team task assignment',
-      'Team calendar view'
-    ]
-  }
-};
-
-/**
- * Generate test file content for a suite
- */
-function generateTestFile(suiteName, config) {
-  const { name, route, tests } = config;
-
-  return `import { test, expect } from '@playwright/test';
-
-const BASE_URL = process.env.BASE_URL || '${BASE_URL}';
-
-test.describe('EHG ${name} Tests', () => {
-  test.beforeEach(async ({ page, context }) => {
-    // Clear cookies and login
-    await context.clearCookies();
-    await page.goto(BASE_URL);
-
-    // Login flow (adjust based on actual app)
-    const needsLogin = await page.url().includes('login');
-    if (needsLogin) {
-      await page.fill('#signin-email, input[type="email"]', 'test@example.com');
-      await page.fill('#signin-password, input[type="password"]', 'Test123!');
-      await page.click('button:has-text("Sign In")');
-      await page.waitForURL(/chairman|dashboard/, { timeout: 10000 });
-    }
-
-    // Navigate to test page
-    await page.goto(\`\${BASE_URL}${route}\`);
-    await page.waitForLoadState('networkidle');
-  });
-
-${tests.map((testName, index) => `
-  test('US-UAT-${suiteName.toUpperCase()}-${String(index + 1).padStart(3, '0')}: ${testName}', async ({ page }) => {
-    // ${testName} implementation
-    ${generateTestImplementation(suiteName, testName, index)}
-  });`).join('\n')}
-});
-
-// Helper functions
-async function waitForElement(page, selector, timeout = 5000) {
-  await page.waitForSelector(selector, { timeout });
+// CLI execution - delegate to the domain module
+if (import.meta.url === `file://${process.argv[1]}`) {
+  const { generateAllTests } = await import('./generate-uat/index.js');
+  generateAllTests();
 }
-
-async function clickAndWait(page, selector, waitForUrl = null) {
-  await page.click(selector);
-  if (waitForUrl) {
-    await page.waitForURL(waitForUrl, { timeout: 5000 });
-  } else {
-    await page.waitForTimeout(1000);
-  }
-}
-
-async function fillForm(page, formData) {
-  for (const [selector, value] of Object.entries(formData)) {
-    await page.fill(selector, value);
-  }
-}
-
-async function verifyToast(page, message) {
-  const toast = page.locator(\`text=/\${message}/i\`);
-  await expect(toast).toBeVisible({ timeout: 5000 });
-}
-
-async function takeScreenshot(page, name) {
-  await page.screenshot({
-    path: \`test-results/screenshots/${suiteName}-\${name}-\${Date.now()}.png\`,
-    fullPage: true
-  });
-}
-
-async function checkAccessibility(page) {
-  // Basic accessibility checks
-  const images = await page.$$('img:not([alt])');
-  expect(images.length).toBe(0);
-
-  const buttons = await page.$$('button:not([aria-label]):not(:has-text(*))');
-  expect(buttons.length).toBe(0);
-}
-
-async function measurePerformance(page, actionName) {
-  const startTime = Date.now();
-  // Action would be performed here
-  const endTime = Date.now();
-  const duration = endTime - startTime;
-
-  console.log(\`Performance: \${actionName} took \${duration}ms\`);
-  expect(duration).toBeLessThan(3000); // 3 second max
-}
-`;
-}
-
-/**
- * Generate specific test implementation based on test name
- */
-function generateTestImplementation(suiteName, testName, index) {
-  const testLower = testName.toLowerCase();
-
-  // Generate contextual test implementation
-  if (testLower.includes('load')) {
-    return `
-    // Verify page loads successfully
-    await expect(page).toHaveURL(new RegExp('${suiteName}'));
-    const pageTitle = await page.title();
-    expect(pageTitle).toBeTruthy();
-
-    // Check for no console errors
-    const errors = [];
-    page.on('console', msg => {
-      if (msg.type() === 'error') errors.push(msg.text());
-    });
-    await page.waitForTimeout(1000);
-    expect(errors).toHaveLength(0);
-
-    // Verify main content is visible
-    const mainContent = page.locator('main, [role="main"], .main-content').first();
-    await expect(mainContent).toBeVisible();`;
-  }
-
-  if (testLower.includes('create') || testLower.includes('add')) {
-    return `
-    // Click create/add button
-    const createBtn = page.locator('button:has-text("Create"), button:has-text("Add"), button:has-text("New")').first();
-    await expect(createBtn).toBeVisible();
-    await createBtn.click();
-
-    // Wait for form/modal
-    await page.waitForTimeout(1000);
-
-    // Fill in required fields
-    const nameField = page.locator('input[name="name"], input[placeholder*="name" i]').first();
-    if (await nameField.count() > 0) {
-      await nameField.fill(\`Test \${suiteName} \${Date.now()}\`);
-    }
-
-    // Submit form
-    const submitBtn = page.locator('button[type="submit"], button:has-text("Save"), button:has-text("Create")').last();
-    await submitBtn.click();
-
-    // Verify success
-    await page.waitForTimeout(2000);
-    const successMsg = page.locator('text=/success|created|saved/i');
-    await expect(successMsg.first()).toBeVisible({ timeout: 5000 });`;
-  }
-
-  if (testLower.includes('edit') || testLower.includes('update')) {
-    return `
-    // Find and click edit button
-    const editBtn = page.locator('button:has-text("Edit"), [aria-label*="edit" i]').first();
-    await expect(editBtn).toBeVisible();
-    await editBtn.click();
-
-    // Wait for edit form
-    await page.waitForTimeout(1000);
-
-    // Update fields
-    const inputField = page.locator('input[type="text"]').first();
-    if (await inputField.count() > 0) {
-      await inputField.clear();
-      await inputField.fill(\`Updated \${Date.now()}\`);
-    }
-
-    // Save changes
-    const saveBtn = page.locator('button:has-text("Save"), button:has-text("Update")').first();
-    await saveBtn.click();
-
-    // Verify update
-    await page.waitForTimeout(1000);
-    await expect(page.locator('text=/updated|saved/i').first()).toBeVisible();`;
-  }
-
-  if (testLower.includes('delete') || testLower.includes('remove')) {
-    return `
-    // Find and click delete button
-    const deleteBtn = page.locator('button:has-text("Delete"), [aria-label*="delete" i]').first();
-    await expect(deleteBtn).toBeVisible();
-    await deleteBtn.click();
-
-    // Handle confirmation dialog
-    const confirmBtn = page.locator('button:has-text("Confirm"), button:has-text("Yes")').first();
-    if (await confirmBtn.count() > 0) {
-      await confirmBtn.click();
-    }
-
-    // Verify deletion
-    await page.waitForTimeout(1000);
-    const successMsg = page.locator('text=/deleted|removed/i');
-    await expect(successMsg.first()).toBeVisible({ timeout: 5000 });`;
-  }
-
-  if (testLower.includes('search') || testLower.includes('filter')) {
-    return `
-    // Find search/filter input
-    const searchInput = page.locator('input[type="search"], input[placeholder*="search" i]').first();
-    await expect(searchInput).toBeVisible();
-
-    // Enter search term
-    await searchInput.fill('test search ${index}');
-    await page.keyboard.press('Enter');
-
-    // Wait for results
-    await page.waitForTimeout(1000);
-
-    // Verify filtered results
-    const results = page.locator('[role="list"] > *, tbody tr, .card, .item');
-    const count = await results.count();
-    console.log(\`Found \${count} filtered results\`);
-
-    // Clear filter
-    const clearBtn = page.locator('button:has-text("Clear"), button[aria-label*="clear" i]').first();
-    if (await clearBtn.count() > 0) {
-      await clearBtn.click();
-    }`;
-  }
-
-  if (testLower.includes('export') || testLower.includes('download')) {
-    return `
-    // Find and click export button
-    const exportBtn = page.locator('button:has-text("Export"), button:has-text("Download")').first();
-    await expect(exportBtn).toBeVisible();
-
-    // Start download promise
-    const downloadPromise = page.waitForEvent('download');
-    await exportBtn.click();
-
-    // Handle export options if present
-    const pdfOption = page.locator('text="PDF"').first();
-    if (await pdfOption.count() > 0) {
-      await pdfOption.click();
-    }
-
-    // Wait for download
-    const download = await downloadPromise;
-    expect(download).toBeTruthy();
-    console.log('Download completed:', await download.suggestedFilename());`;
-  }
-
-  if (testLower.includes('sort')) {
-    return `
-    // Find sortable column header
-    const sortHeader = page.locator('th, [role="columnheader"]').first();
-    await expect(sortHeader).toBeVisible();
-
-    // Click to sort
-    await sortHeader.click();
-    await page.waitForTimeout(500);
-
-    // Verify sort indicator
-    const sortIcon = sortHeader.locator('[aria-label*="sort" i], .sort-icon');
-    await expect(sortIcon.first()).toBeVisible();
-
-    // Click again for reverse sort
-    await sortHeader.click();
-    await page.waitForTimeout(500);`;
-  }
-
-  if (testLower.includes('permission') || testLower.includes('access')) {
-    return `
-    // Check for permission-based elements
-    const adminOnly = page.locator('[data-permission="admin"], .admin-only');
-    const adminCount = await adminOnly.count();
-
-    // Verify based on user role
-    const userRole = 'user'; // This would be dynamic
-    if (userRole === 'admin') {
-      expect(adminCount).toBeGreaterThan(0);
-    } else {
-      expect(adminCount).toBe(0);
-    }
-
-    // Try accessing restricted action
-    const restrictedBtn = page.locator('button:has-text("Admin")').first();
-    if (await restrictedBtn.count() > 0) {
-      const isDisabled = await restrictedBtn.isDisabled();
-      expect(isDisabled).toBe(userRole !== 'admin');
-    }`;
-  }
-
-  if (testLower.includes('performance') || testLower.includes('metric')) {
-    return `
-    // Measure page performance
-    const metrics = await page.evaluate(() => {
-      const perf = performance.getEntriesByType('navigation')[0];
-      return {
-        domContentLoaded: perf.domContentLoadedEventEnd - perf.domContentLoadedEventStart,
-        loadComplete: perf.loadEventEnd - perf.loadEventStart,
-        firstPaint: performance.getEntriesByName('first-paint')[0]?.startTime || 0
-      };
-    });
-
-    console.log('Performance metrics:', metrics);
-    expect(metrics.loadComplete).toBeLessThan(3000);
-
-    // Check for performance indicators on page
-    const perfWidget = page.locator('[data-testid*="performance"], .performance-metric').first();
-    if (await perfWidget.count() > 0) {
-      await expect(perfWidget).toBeVisible();
-    }`;
-  }
-
-  // Default implementation for other test types
-  return `
-    // Generic test implementation for: ${testName}
-    await page.waitForLoadState('networkidle');
-
-    // Take screenshot for visual verification
-    await takeScreenshot(page, '${testName.replace(/\s+/g, '-').toLowerCase()}');
-
-    // Basic interaction test
-    const interactiveElements = page.locator('button, a, input, select');
-    const elementCount = await interactiveElements.count();
-    expect(elementCount).toBeGreaterThan(0);
-
-    // Check for expected content
-    const pageContent = await page.content();
-    expect(pageContent).toBeTruthy();
-
-    console.log('Test ${index + 1} completed: ${testName}');`;
-}
-
-/**
- * Generate administrative function tests
- */
-function generateAdminTests() {
-  const adminSuites = {
-    settings: {
-      name: 'System Settings',
-      route: '/settings',
-      tests: [
-        'Settings page loads',
-        'General settings section',
-        'Update company information',
-        'Change timezone settings',
-        'Update notification preferences',
-        'Email configuration',
-        'API settings management',
-        'Integration toggles',
-        'Theme customization',
-        'Language selection',
-        'Save settings changes',
-        'Reset to defaults',
-        'Export settings',
-        'Import settings',
-        'Settings validation'
-      ]
-    },
-    security: {
-      name: 'Security Management',
-      route: '/security',
-      tests: [
-        'Security dashboard loads',
-        'View security policies',
-        'Two-factor authentication setup',
-        'Password requirements configuration',
-        'Session timeout settings',
-        'IP whitelist management',
-        'Security audit log',
-        'Vulnerability scan results',
-        'Security notifications',
-        'Access control lists',
-        'API key rotation',
-        'Certificate management',
-        'Security report generation',
-        'Incident response tools',
-        'Security training modules'
-      ]
-    }
-  };
-
-  return adminSuites;
-}
-
-/**
- * Generate cross-functional tests
- */
-function generateCrossFunctionalTests() {
-  const crossFunctionalSuites = {
-    accessibility: {
-      name: 'Accessibility Compliance',
-      route: '/',
-      tests: [
-        'Keyboard navigation throughout app',
-        'Screen reader compatibility',
-        'ARIA labels present',
-        'Color contrast compliance',
-        'Focus indicators visible',
-        'Skip navigation links',
-        'Form field labels',
-        'Error message association',
-        'Alternative text for images',
-        'Semantic HTML structure',
-        'Heading hierarchy',
-        'Table headers association',
-        'Link purpose clarity',
-        'Time limits adjustable',
-        'Motion control options'
-      ]
-    },
-    performance: {
-      name: 'Performance Metrics',
-      route: '/',
-      tests: [
-        'Page load time under 3 seconds',
-        'Time to interactive measurement',
-        'First contentful paint',
-        'Largest contentful paint',
-        'Cumulative layout shift',
-        'API response times',
-        'Database query optimization',
-        'Image optimization check',
-        'Bundle size validation',
-        'Memory leak detection'
-      ]
-    },
-    mobile: {
-      name: 'Mobile Responsive',
-      route: '/',
-      tests: [
-        'Mobile viewport rendering',
-        'Touch gesture support',
-        'Swipe navigation',
-        'Mobile menu functionality',
-        'Form input on mobile',
-        'Mobile table display',
-        'Image scaling on mobile',
-        'Mobile performance metrics',
-        'Offline functionality',
-        'Progressive web app features'
-      ]
-    }
-  };
-
-  return crossFunctionalSuites;
-}
-
-/**
- * Generate end-to-end scenario tests
- */
-function generateE2ETests() {
-  const e2eSuites = {
-    ventureLifecycle: {
-      name: 'Complete Venture Lifecycle',
-      route: '/ventures',
-      tests: [
-        'Create new venture from scratch',
-        'Add complete venture details',
-        'Assign team members',
-        'Set milestones and KPIs',
-        'Generate initial report',
-        'Update progress metrics',
-        'Trigger status change',
-        'Request approval workflow',
-        'Receive approval',
-        'Archive completed venture'
-      ]
-    },
-    executiveReporting: {
-      name: 'Executive Reporting Flow',
-      route: '/analytics',
-      tests: [
-        'Access analytics dashboard',
-        'Select report template',
-        'Configure report parameters',
-        'Add custom metrics',
-        'Generate preview',
-        'Review and edit',
-        'Export to multiple formats',
-        'Share with stakeholders',
-        'Schedule recurring report',
-        'Verify delivery'
-      ]
-    }
-  };
-
-  return e2eSuites;
-}
-
-/**
- * Main function to generate all test files
- */
-async function generateAllTests() {
-  console.log(chalk.bold.cyan(`
-╔══════════════════════════════════════════════════════════════╗
-║     Comprehensive UAT Test Suite Generator for EHG           ║
-╚══════════════════════════════════════════════════════════════╝
-  `));
-
-  try {
-    // Ensure test directory exists
-    await fs.mkdir(TEST_DIR, { recursive: true });
-
-    let totalTests = 0;
-    let filesGenerated = 0;
-
-    // Phase 1: Core User Journey Tests
-    console.log(chalk.blue('\n📝 Phase 1: Generating Core User Journey Tests...'));
-    for (const [suiteName, config] of Object.entries(TEST_SUITES)) {
-      const fileName = `${suiteName}.spec.js`;
-      const filePath = path.join(TEST_DIR, fileName);
-      const content = generateTestFile(suiteName, config);
-
-      await fs.writeFile(filePath, content);
-      filesGenerated++;
-      totalTests += config.tests.length;
-
-      console.log(chalk.green(`   ✓ Generated ${fileName} (${config.tests.length} tests)`));
-    }
-
-    // Phase 2: Administrative Function Tests
-    console.log(chalk.blue('\n📝 Phase 2: Generating Administrative Tests...'));
-    const adminSuites = generateAdminTests();
-    for (const [suiteName, config] of Object.entries(adminSuites)) {
-      const fileName = `${suiteName}.spec.js`;
-      const filePath = path.join(TEST_DIR, fileName);
-      const content = generateTestFile(suiteName, config);
-
-      await fs.writeFile(filePath, content);
-      filesGenerated++;
-      totalTests += config.tests.length;
-
-      console.log(chalk.green(`   ✓ Generated ${fileName} (${config.tests.length} tests)`));
-    }
-
-    // Phase 3: Cross-Functional Tests
-    console.log(chalk.blue('\n📝 Phase 3: Generating Cross-Functional Tests...'));
-    const crossSuites = generateCrossFunctionalTests();
-    for (const [suiteName, config] of Object.entries(crossSuites)) {
-      const fileName = `${suiteName}.spec.js`;
-      const filePath = path.join(TEST_DIR, fileName);
-      const content = generateTestFile(suiteName, config);
-
-      await fs.writeFile(filePath, content);
-      filesGenerated++;
-      totalTests += config.tests.length;
-
-      console.log(chalk.green(`   ✓ Generated ${fileName} (${config.tests.length} tests)`));
-    }
-
-    // Phase 4: End-to-End Tests
-    console.log(chalk.blue('\n📝 Phase 4: Generating End-to-End Scenario Tests...'));
-    const e2eSuites = generateE2ETests();
-    for (const [suiteName, config] of Object.entries(e2eSuites)) {
-      const fileName = `e2e-${suiteName}.spec.js`;
-      const filePath = path.join(TEST_DIR, fileName);
-      const content = generateTestFile(suiteName, config);
-
-      await fs.writeFile(filePath, content);
-      filesGenerated++;
-      totalTests += config.tests.length;
-
-      console.log(chalk.green(`   ✓ Generated ${fileName} (${config.tests.length} tests)`));
-    }
-
-    // Generate test runner script
-    console.log(chalk.blue('\n📝 Generating Test Runner Script...'));
-    await generateTestRunner();
-
-    // Summary
-    console.log(chalk.bold.green(`
-╔══════════════════════════════════════════════════════════════╗
-║                    Generation Complete!                       ║
-╚══════════════════════════════════════════════════════════════╝
-
-📊 Statistics:
-   • Test Files Generated: ${filesGenerated}
-   • Total Test Cases: ${totalTests}
-   • Coverage: 43+ pages
-   • Test Categories: 4 phases
-
-📁 Location: ${TEST_DIR}
-
-🚀 Run Tests:
-   • All tests: npm run test:uat
-   • Specific suite: npm run test:uat -- ventures
-   • With UI: npm run test:uat:ui
-
-📈 Expected Coverage:
-   • UI Coverage: >95%
-   • User Journeys: 100%
-   • Critical Paths: 100%
-   • Edge Cases: ~80%
-    `));
-
-  } catch (error) {
-    console.error(chalk.red('❌ Error generating tests:'), error);
-    process.exit(1);
-  }
-}
-
-/**
- * Generate test runner script
- */
-async function generateTestRunner() {
-  const _runnerContent = `{
-  "scripts": {
-    "test:uat": "playwright test tests/uat --config=playwright-uat.config.js",
-    "test:uat:ui": "playwright test tests/uat --config=playwright-uat.config.js --ui",
-    "test:uat:debug": "playwright test tests/uat --config=playwright-uat.config.js --debug",
-    "test:uat:report": "playwright show-report",
-    "test:uat:coverage": "node scripts/calculate-coverage.js"
-  }
-}`;
-
-  const packagePath = path.join(__dirname, '..', 'package.json');
-  const packageJson = JSON.parse(await fs.readFile(packagePath, 'utf-8'));
-
-  // Add UAT test scripts
-  packageJson.scripts = {
-    ...packageJson.scripts,
-    'test:uat': 'playwright test tests/uat --config=playwright-uat.config.js',
-    'test:uat:ui': 'playwright test tests/uat --config=playwright-uat.config.js --ui',
-    'test:uat:debug': 'playwright test tests/uat --config=playwright-uat.config.js --debug',
-    'test:uat:report': 'playwright show-report',
-    'test:uat:coverage': 'node scripts/calculate-coverage.js'
-  };
-
-  await fs.writeFile(packagePath, JSON.stringify(packageJson, null, 2));
-  console.log(chalk.green('   ✓ Updated package.json with test scripts'));
-}
-
-// Execute the generator
-generateAllTests();
