+++ scripts/generate-uat/suite-builders.js
@@ -1,0 +1,166 @@
+/**
+ * Suite Builders Domain
+ * Generates admin, cross-functional, and E2E test suites
+ *
+ * @module generate-uat/suite-builders
+ */
+
+/**
+ * Generate administrative function tests
+ * @returns {Object} Admin test suites configuration
+ */
+export function generateAdminTests() {
+  return {
+    settings: {
+      name: 'System Settings',
+      route: '/settings',
+      tests: [
+        'Settings page loads',
+        'General settings section',
+        'Update company information',
+        'Change timezone settings',
+        'Update notification preferences',
+        'Email configuration',
+        'API settings management',
+        'Integration toggles',
+        'Theme customization',
+        'Language selection',
+        'Save settings changes',
+        'Reset to defaults',
+        'Export settings',
+        'Import settings',
+        'Settings validation'
+      ]
+    },
+    security: {
+      name: 'Security Management',
+      route: '/security',
+      tests: [
+        'Security dashboard loads',
+        'View security policies',
+        'Two-factor authentication setup',
+        'Password requirements configuration',
+        'Session timeout settings',
+        'IP whitelist management',
+        'Security audit log',
+        'Vulnerability scan results',
+        'Security notifications',
+        'Access control lists',
+        'API key rotation',
+        'Certificate management',
+        'Security report generation',
+        'Incident response tools',
+        'Security training modules'
+      ]
+    }
+  };
+}
+
+/**
+ * Generate cross-functional tests
+ * @returns {Object} Cross-functional test suites configuration
+ */
+export function generateCrossFunctionalTests() {
+  return {
+    accessibility: {
+      name: 'Accessibility Compliance',
+      route: '/',
+      tests: [
+        'Keyboard navigation throughout app',
+        'Screen reader compatibility',
+        'ARIA labels present',
+        'Color contrast compliance',
+        'Focus indicators visible',
+        'Skip navigation links',
+        'Form field labels',
+        'Error message association',
+        'Alternative text for images',
+        'Semantic HTML structure',
+        'Heading hierarchy',
+        'Table headers association',
+        'Link purpose clarity',
+        'Time limits adjustable',
+        'Motion control options'
+      ]
+    },
+    performance: {
+      name: 'Performance Metrics',
+      route: '/',
+      tests: [
+        'Page load time under 3 seconds',
+        'Time to interactive measurement',
+        'First contentful paint',
+        'Largest contentful paint',
+        'Cumulative layout shift',
+        'API response times',
+        'Database query optimization',
+        'Image optimization check',
+        'Bundle size validation',
+        'Memory leak detection'
+      ]
+    },
+    mobile: {
+      name: 'Mobile Responsive',
+      route: '/',
+      tests: [
+        'Mobile viewport rendering',
+        'Touch gesture support',
+        'Swipe navigation',
+        'Mobile menu functionality',
+        'Form input on mobile',
+        'Mobile table display',
+        'Image scaling on mobile',
+        'Mobile performance metrics',
+        'Offline functionality',
+        'Progressive web app features'
+      ]
+    }
+  };
+}
+
+/**
+ * Generate end-to-end scenario tests
+ * @returns {Object} E2E test suites configuration
+ */
+export function generateE2ETests() {
+  return {
+    ventureLifecycle: {
+      name: 'Complete Venture Lifecycle',
+      route: '/ventures',
+      tests: [
+        'Create new venture from scratch',
+        'Add complete venture details',
+        'Assign team members',
+        'Set milestones and KPIs',
+        'Generate initial report',
+        'Update progress metrics',
+        'Trigger status change',
+        'Request approval workflow',
+        'Receive approval',
+        'Archive completed venture'
+      ]
+    },
+    executiveReporting: {
+      name: 'Executive Reporting Flow',
+      route: '/analytics',
+      tests: [
+        'Access analytics dashboard',
+        'Select report template',
+        'Configure report parameters',
+        'Add custom metrics',
+        'Generate preview',
+        'Review and edit',
+        'Export to multiple formats',
+        'Share with stakeholders',
+        'Schedule recurring report',
+        'Verify delivery'
+      ]
+    }
+  };
+}
+
+export default {
+  generateAdminTests,
+  generateCrossFunctionalTests,
+  generateE2ETests
+};
